chore: move `HomeView` to a new location
