--- lib/views/home_view.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomeView extends StatelessWidget {
-  const HomeView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Home'),
-      ),
-      body: const Center(
-        child: Text('Home View'),
-      ),
-    );
-  }
-}
